Small path meta tweaks
